test: move test orders to fixture
