+++ tests/unit/store/actions.spec.ts
@@ -2,11 +2,11 @@ import actions from '@/store/actions'
 import router from '@/router'
 import { Cart } from '@/types/Cart'
 import emptyState from '@/store/state'
+import orders from '../fixtures/orders'
 
 // @ts-ignore
 import { post } from '@molgenis/molgenis-api-client'
 import ApplicationState from '@/types/ApplicationState'
-import { Order, OrderState } from '@/types/Order'
 
 const cart: Cart = {
   selection: [{
@@ -18,21 +18,6 @@ const cart: Cart = {
   }
 }
 const cartContents = JSON.stringify(cart)
-
-const orders: Order[] = [{
-  id: 'edcba',
-  state: OrderState.Draft,
-  submissionDate: '2019-10-31T13:48:12Z'
-}, {
-  id: 'abcde',
-  state: OrderState.Submitted,
-  submissionDate: '2019-10-30T13:25:49Z',
-  name: 'My draft order',
-  applicationForm: {
-    filename: 'Motivation.pdf',
-    id: 'aaaac3rcetmgfmudsodb3laaay',
-    url: 'https://lifelines.test.molgenis.org/files/aaaac3rcetmgfmudsodb3laaay'
-  } }]
 
 const mockResponses: {[key:string]: Object} = {
   '/api/v2/lifelines_cart?num=10000': {
